Add a rest service for requests concerning members and resolve minor code smells
