+++ src/main/java/de/hpi/bpt/chimera/rest/filters/AuthorizationRequestFilter.java
@@ -29,9 +29,9 @@ import de.hpi.bpt.chimera.usermanagment.UserManager;
 import java.io.IOException;
 import java.util.List;
 
-/**
- *
- */
+// TODO: extend this Provider. 
+// Either by using 'getMatchedResources' to look for the called RestService and not the id.
+// Or create a costume request filter for every RestService via static or dynamic 'NameBinding'.
 @Provider
 public class AuthorizationRequestFilter implements ContainerRequestFilter {
 	private static Logger log = Logger.getLogger(AuthorizationRequestFilter.class);
@@ -47,7 +47,7 @@ public class AuthorizationRequestFilter implements ContainerRequestFilter {
 	@Override
 	public void filter(ContainerRequestContext requestContext) throws IOException {
 		this.requestContext = requestContext;
-		log.info("Filter");
+
 		try {
 			validateUser();
 			validateOrganization();
@@ -72,117 +72,13 @@ public class AuthorizationRequestFilter implements ContainerRequestFilter {
 			Response unauthorized = Response.status(Response.Status.UNAUTHORIZED).type(MediaType.APPLICATION_JSON).entity(message.toString()).build();
 			this.requestContext.abortWith(unauthorized);
 		}
-		log.info("Filter End");
-		// if (!isValidScenario()) {
-		// abortIllegalScenario(requestContext.getMethod());
-		// } else if (!isValidInstance()) {
-		// abortInvalidInstance(requestContext.getMethod());
-		// } else if (!isValidActivity()) {
-		// abortIllegalActivity(requestContext.getMethod());
-		// } else if (!isValidActivityInstance()) {
-		// abortIllegalActivityInstance(requestContext.getMethod());
-		// }
 	}
-
-	/*
-	private void abortIllegalActivity(String method) {
-		int activityId = Integer.parseInt(requestContext.getUriInfo().getPathParameters().getFirst("activityId"));
-		if (method.equals(HttpMethod.GET.toString())) {
-			this.abortWithNotFound(String.format(errorMsg, "activity", activityId));
-		} else {
-			this.abortBadRequest(String.format(errorMsg, "activity", activityId));
-		}
-	}
-
-	private void abortIllegalActivityInstance(String method) {
-		int activityInstanceId = Integer.parseInt(requestContext.getUriInfo().getPathParameters().getFirst("activityInstanceId"));
-		if (method.equals(HttpMethod.GET.toString())) {
-			this.abortWithNotFound(String.format(errorMsg, "activity instance", activityInstanceId));
-		} else {
-			this.abortBadRequest(String.format(errorMsg, "activity instance", activityInstanceId));
-		}
-	}
-
-	private boolean isValidActivityInstance() {
-		MultivaluedMap<String, String> map = requestContext.getUriInfo().getPathParameters();
-		if (!map.containsKey("activityInstanceId")) {
-			return true;
-		}
-		assert map.containsKey("instanceId") && map.containsKey("scenarioId") : "Activity Id is only defined in context of a scenario instance";
-		int instanceId = Integer.parseInt(map.getFirst("instanceId"));
-		int activityInstanceId = Integer.parseInt(map.getFirst("activityInstanceId"));
-
-		return new DbActivityInstance().existsActivityInstance(activityInstanceId, instanceId);
-	}
-
-	private boolean isValidActivity() {
-		MultivaluedMap<String, String> map = requestContext.getUriInfo().getPathParameters();
-		if (!map.containsKey("activityId")) {
-			return true;
-		}
-		assert map.containsKey("instanceId") && map.containsKey("scenarioId") : "Activity Id is only defined in context of a scenario instance";
-		int scenarioId = Integer.parseInt(map.getFirst("scenarioId"));
-		int activityId = Integer.parseInt(map.getFirst("activityId"));
-
-		return new DbControlNode().existControlNode(activityId, scenarioId);
-	}
-
-	private void abortIllegalScenario(String method) {
-		int scenarioId = Integer.parseInt(requestContext.getUriInfo().getPathParameters().getFirst("scenarioId"));
-		if (method.equals(HttpMethod.GET.toString())) {
-			this.abortWithNotFound(String.format(errorMsg, "scenario", scenarioId));
-		} else {
-			this.abortBadRequest(String.format(errorMsg, "scenario", scenarioId));
-		}
-	}
-
-	private boolean isValidScenario() {
-		/*
-		 * MultivaluedMap<String, String> map =
-		 * requestContext.getUriInfo().getPathParameters(); if
-		 * (!map.containsKey("scenarioId")) { return true; } int scenarioId =
-		 * Integer.parseInt(map.getFirst("scenarioId")); return new
-		 * DbScenario().existScenario(scenarioId);
-		 *//*
-		return true;
-	}
-
-	private boolean isValidInstance() {
-		MultivaluedMap<String, String> map = requestContext.getUriInfo().getPathParameters();
-		if (!map.containsKey("instanceId")) {
-			return true;
-		}
-
-		int scenarioId = Integer.parseInt(map.getFirst("scenarioId"));
-		int instanceId = Integer.parseInt(map.getFirst("instanceId"));
-		return new DbScenarioInstance().doesScenarioInstanceBelongToScenario(scenarioId, instanceId);
-	}
-
-	private void abortInvalidInstance(String method) {
-		int instanceId = Integer.parseInt(requestContext.getUriInfo().getPathParameters().getFirst("instanceId"));
-
-		if (method.equals(HttpMethod.GET.toString())) {
-			abortWithNotFound(String.format(errorMsg, "scenario instance", instanceId));
-		} else {
-			abortBadRequest(String.format(errorMsg, "scenario instance", instanceId));
-		}
-	}
-
-	private void abortWithNotFound(String errorMsg) {
-		Response notFound = Response.status(Response.Status.NOT_FOUND).type(MediaType.APPLICATION_JSON).entity(errorMsg).build();
-		this.requestContext.abortWith(notFound);
-	}
-
-	private void abortBadRequest(String errorMsg) {
-		Response badRequest = Response.status(Response.Status.BAD_REQUEST).type(MediaType.APPLICATION_JSON).entity(errorMsg).build();
-		this.requestContext.abortWith(badRequest);
-	}
-	*/
 
 	private void validateUser() {
 		// TODO: think about costum response message for WebApplicationException
 		String method = requestContext.getMethod();
 		String path = requestContext.getUriInfo().getPath();
+		// every request is allowed to create a new user
 		if ("POST".equals(method) && path.matches("(v3/users)(/)?")) {
 			return;
 		}
@@ -227,7 +123,7 @@ public class AuthorizationRequestFilter implements ContainerRequestFilter {
 		String orgId = parameters.getFirst("organizationId");
 		try {
 			organization = OrganizationManager.getOrganizationById(orgId);
-			if (!organization.isMember(requester)) {
+			if (!organization.isMember(requester) || !requester.isAdmin()) {
 				throw new IllegalArgumentException(UNAUTHORIZED_MEMBER_MESSAGE);
 			}
 		} catch (IllegalOrganizationIdException e) {
@@ -278,7 +174,6 @@ public class AuthorizationRequestFilter implements ContainerRequestFilter {
 			throw e;
 		}
 	}
-
 
 	private void validateActivityInstance() {
 		MultivaluedMap<String, String> parameters = requestContext.getUriInfo().getPathParameters();
